requirements.txt
# ==================== WEB FRAMEWORK ====================
streamlit==1.28.1

# ==================== COMPUTER VISION ====================
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0

# ==================== FACE RECOGNITION ====================
insightface==0.7.3
onnxruntime==1.16.1
mediapipe==0.10.3
retinaface==0.0.13
scikit-image==0.21.0
scipy==1.11.2

# ==================== DEEP LEARNING ====================
torch==2.0.1
torchvision==0.15.2

# ==================== DATA PROCESSING ====================
pandas==2.0.3
sqlalchemy==2.0.21

# ==================== CONFIGURATION ====================
python-dotenv==1.0.0

# ==================== API & NETWORKING ====================
requests==2.31.0

# ==================== DATABASE ====================
sqlite3-python==1.0.0
